fix: Notice
1. 공지사항 PK 자료형 수정
2. db 수정 요망(Auto_Increment)

+++ Backend/erpc/src/main/java/com/cineverse/erpc/notice/board/aggregate/entity/NoticeBoard.java
@@ -17,7 +17,7 @@ public class NoticeBoard {
     @Id
     @Column(name = "notice_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int noticeId;
+    private Long noticeId;
 
     @Column(name = "notice_title")
     private String noticeTitle;
