+++ Backend/erpc/src/main/java/com/cineverse/erpc/notice/board/repository/NoticeBoardRepository.java
@@ -3,5 +3,5 @@ package com.cineverse.erpc.notice.board.repository;
 import com.cineverse.erpc.notice.board.aggregate.entity.NoticeBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeBoardRepository extends JpaRepository<NoticeBoard, Integer> {
+public interface NoticeBoardRepository extends JpaRepository<NoticeBoard, Long> {
 }
